jazzy

+++ src/tools/point_types.h
@@ -11,8 +11,6 @@
 
 #include "common/eigen_types.h"
 
-#include <livox_ros_driver/CustomMsg.h>
-
 namespace zjloc
 {
 
@@ -140,7 +138,7 @@ namespace ouster_ros
         float intensity;
         uint32_t t;
         uint16_t reflectivity;
-        uint8_t ring;
+        uint16_t ring;
         uint16_t ambient;
         uint32_t range;
         EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -156,7 +154,7 @@ POINT_CLOUD_REGISTER_POINT_STRUCT(ouster_ros::Point,
                                       // use std::uint32_t to avoid conflicting with pcl::uint32_t
                                       (std::uint32_t, t, t)
                                       (std::uint16_t, reflectivity, reflectivity)
-                                      (std::uint8_t, ring, ring)
+                                      (std::uint16_t, ring, ring)
                                       (std::uint16_t, ambient, ambient)
                                       (std::uint32_t, range, range)
 )
